refactor: BaseApplication 삭제 및 BaseActivity 수정
