+++ 박수진/week1/.idea/deploymentTargetDropDown.xml
@@ -12,6 +12,6 @@
         </deviceKey>
       </Target>
     </targetSelectedWithDropDown>
-    <timeTargetWasSelectedWithDropDown value="2022-04-27T15:41:25.027773700Z" />
+    <timeTargetWasSelectedWithDropDown value="2022-04-28T08:15:01.995449100Z" />
   </component>
 </project>
