+++ 박수진/week1/app/src/main/res/layout/activity_main.xml
@@ -38,6 +38,7 @@
         app:layout_constraintEnd_toEndOf="parent"
         app:layout_constraintStart_toStartOf="parent"
         app:layout_constraintTop_toBottomOf="@+id/search_bar"
+        app:layoutManager="androidx.recyclerview.widget.LinearLayoutManager"
         android:padding="6dp"/>
 
 </androidx.constraintlayout.widget.ConstraintLayout>
